Uses sprout-git::projects instead of sprout-bosh::repos
We don't want to maintain our own custom recipe for installing git
projects so we are switching to a recipe inside the
opensource sprout-git cookbook.

No longer install the following repos as we don't actively use them for
development:
 - bosh-checkman
 - bosh-jenkins-config
 - bosh-sample-release
 - bosh-packer-templates
 - bosh-ci

Added projects for the 4 cpi's that we maintain
 - bosh-aws-cpi-release
 - bosh-openstack-cpi-release
 - bosh-vcloud-cpi-release
 - bosh-vsphere-cpi-release

runs `git submodule update --init --recursive` for each repo we clone so that the developers don't have to.

Note: we don't run `bundle install` for the repos as soloist is run usually
using the System version of ruby and the projects are usually maintained in a
managed version of ruby so there is no need to have the gems installed in the
System ruby.  For now developers will have to run the bundle command in the
ruby version of their choosing.

[#100074508](https://www.pivotaltracker.com/story/show/100074508)

Signed-off-by: Rexxar Liang (Biao Liang) <rexxar.liang@huawei.com>